Version 4.0.2
*  config/configure.ac configure include/version.inc interface/init_load.F linear_algebra/SLK_POOLS_init.F linear_algebra/SLK_test.F modules/mod_SLK.F parallel/PARALLEL_defaults.F parallel/PARALLEL_get_user_structure.F parallel/PARALLEL_global_indexes.F 

SLK now fully working also reading the PARALLEL conf from input. 

Still POOLS must be created by the SLK_POOLS_init as I was not able to transfer in the POOL definition the
COMM_type. This is still to be done.

Also the code get frozen if the BLACS grid does not involve all CPU's of the POOL.

Patch sent by: Andrea


git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/yambo-devel/branches/devel-andrea@4771 76a4dfbf-5f1f-0410-a543-a0282073b526